Add files via upload
The modifications are aimed to provide access via SmartPort to all Flight Modes  as provided/defined by the INAV configurator.
This enables  all  users (MULTIROTOR and FW)  to pick freely  from all  flight modes those which they  need instead of being limited to an arbitrarily defined subset .
The  code has been tested with INAV 1.7.3 and needs simply to replace the original  one. nothing else is required to generate a custom build.
Presently the documentation is limited to the comments included in the code. If interest is confirmed a separate documentation will be provided . Beyond that a lua script to decode tmp1 & tmp2 could possibly follow.
If interest is confirmed  the d
The following Flight Modes have been added to the already  existing ones
1) Home Reset
2) Servo Autotrim
3) Auto Tune
4) Blackbox
5) Nav Launch
More can be added since the modified SmartPort.c  uses  an improved  encoding logic which allows  more information to be transported within bandwith provided by the tmp1 container.